+++ essrpc/tests/basic.rs
@@ -11,7 +11,7 @@ use serde_derive::{Deserialize, Serialize};
 
 use essrpc::essrpc;
 use essrpc::transports::{BincodeTransport, JSONTransport};
-use essrpc::{RPCClient, RPCServer};
+use essrpc::{RPCClient, RPCErrorKind, RPCServer};
 
 #[derive(Debug, Deserialize, Serialize)]
 pub struct TestError {
@@ -61,19 +61,13 @@ impl Foo for FooImpl {
 #[test]
 fn basic_bincode() {
     let foo = bincode_foo();
-    match foo.bar("the answer".to_string(), 42) {
-        Ok(result) => assert_eq!("the answer is 42", result),
-        Err(e) => panic!("error: {:?}", e),
-    }
+    client42(&foo);
 }
 
 #[test]
 fn basic_json() {
     let foo = json_foo();
-    match foo.bar("the answer".to_string(), 42) {
-        Ok(result) => assert_eq!("the answer is 42", result),
-        Err(e) => panic!("error: {:?}", e),
-    }
+    client42(&foo);
 }
 
 #[test]
@@ -93,12 +87,68 @@ fn serve_multiple() {
         serve.serve()
     });
     let foo = FooRPCClient::new(BincodeTransport::new(s1));
-    match foo.bar("the answer".to_string(), 42) {
-        Ok(result) => assert_eq!("the answer is 42", result),
-        Err(e) => panic!("error: {:?}", e),
-    }
+    client42(&foo);
     match foo.bar("the answer".to_string(), 43) {
         Ok(result) => assert_eq!("the answer is 43", result),
+        Err(e) => panic!("error: {:?}", e),
+    }
+}
+
+#[test]
+fn serve_single_call_ok_bincode() {
+    // Verify that serve_single_call returns an Ok result
+    let (s1, s2) = UnixStream::pair().unwrap();
+    thread::spawn(move || {
+        let foo = FooRPCClient::new(BincodeTransport::new(s1));
+        client42(&foo);
+    });
+    let mut serve = FooRPCServer::new(FooImpl::new(), BincodeTransport::new(s2));
+    serve.serve_single_call().unwrap()
+}
+
+#[test]
+fn serve_single_call_ok_json() {
+    // Verify that serve_single_call returns an Ok result
+    let (s1, s2) = UnixStream::pair().unwrap();
+    thread::spawn(move || {
+        let foo = FooRPCClient::new(JSONTransport::new(s1));
+        client42(&foo);
+    });
+    let mut serve = FooRPCServer::new(FooImpl::new(), JSONTransport::new(s2));
+    serve.serve_single_call().unwrap()
+}
+
+#[test]
+fn serve_multiple_eof_on_disconnect_bincode() {
+    let (s1, s2) = UnixStream::pair().unwrap();
+    thread::spawn(move || {
+        let foo = FooRPCClient::new(BincodeTransport::new(s1));
+        client42(&foo);
+    });
+    let mut serve = FooRPCServer::new(FooImpl::new(), BincodeTransport::new(s2));
+    match serve.serve() {
+        Ok(_) => panic!("Expected EOF error"),
+        Err(e) => assert_eq!(e.kind, RPCErrorKind::TransportEOF),
+    }
+}
+
+#[test]
+fn serve_multiple_eof_on_disconnect_json() {
+    let (s1, s2) = UnixStream::pair().unwrap();
+    thread::spawn(move || {
+        let foo = FooRPCClient::new(JSONTransport::new(s1));
+        client42(&foo);
+    });
+    let mut serve = FooRPCServer::new(FooImpl::new(), JSONTransport::new(s2));
+    match serve.serve() {
+        Ok(_) => panic!("Expected EOF error"),
+        Err(e) => assert_eq!(e.kind, RPCErrorKind::TransportEOF),
+    }
+}
+
+fn client42<T: Foo>(client: &T) {
+    match client.bar("the answer".to_string(), 42) {
+        Ok(result) => assert_eq!("the answer is 42", result),
         Err(e) => panic!("error: {:?}", e),
     }
 }